+++ opentelemetry-sdk/src/opentelemetry/sdk/_configuration/__init__.py
@@ -45,6 +45,8 @@ from opentelemetry.sdk.environment_variables import (
     OTEL_EXPORTER_OTLP_METRICS_PROTOCOL,
     OTEL_EXPORTER_OTLP_PROTOCOL,
     OTEL_EXPORTER_OTLP_TRACES_PROTOCOL,
+    OTEL_PYTHON_LOG_FORMAT,
+    OTEL_PYTHON_LOG_LEVEL,
     OTEL_TRACES_SAMPLER,
     OTEL_TRACES_SAMPLER_ARG,
 )
@@ -89,6 +91,15 @@ _DEFAULT_ID_GENERATOR = _RANDOM_ID_GENERATOR
 
 _OTEL_SAMPLER_ENTRY_POINT_GROUP = "opentelemetry_traces_sampler"
 
+_OTEL_PYTHON_LOG_LEVEL_BY_NAME = {
+    "notset": logging.NOTSET,
+    "debug": logging.DEBUG,
+    "info": logging.INFO,
+    "warn": logging.WARNING,
+    "warning": logging.WARNING,
+    "error": logging.ERROR,
+}
+
 _logger = logging.getLogger(__name__)
 
 
@@ -131,6 +142,13 @@ def _get_sampler() -> str | None:
 
 def _get_id_generator() -> str:
     return environ.get(OTEL_PYTHON_ID_GENERATOR, _DEFAULT_ID_GENERATOR)
+
+
+def _get_log_level() -> int:
+    return _OTEL_PYTHON_LOG_LEVEL_BY_NAME.get(
+        environ.get(OTEL_PYTHON_LOG_LEVEL, "notset").lower().strip(),
+        logging.NOTSET,
+    )
 
 
 def _get_exporter_entry_point(
@@ -255,11 +273,19 @@ def _init_logging(
     if setup_logging_handler:
         _patch_basic_config()
 
-        # Add OTel handler
-        handler = LoggingHandler(
-            level=logging.NOTSET, logger_provider=provider
-        )
-        logging.getLogger().addHandler(handler)
+        # Log Handler
+        root_logger = logging.getLogger()
+        handler = LoggingHandler(logger_provider=provider)
+        # Log level
+        if OTEL_PYTHON_LOG_LEVEL in environ:
+            handler.setLevel(_get_log_level())
+        # Log format
+        if OTEL_PYTHON_LOG_FORMAT in environ:
+            log_format = environ.get(
+                OTEL_PYTHON_LOG_FORMAT, logging.BASIC_FORMAT
+            )
+            handler.setFormatter(logging.Formatter(log_format))
+        root_logger.addHandler(handler)
 
 
 def _patch_basic_config():